pyproject.toml
[tool.poetry]
name = "thesis-csi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
zenml = "0.68.1"
pandas = "^2.2.2"
pyarrow = "^15.0.0"
scikit-learn = "^1.4.0"
numpy = "^1.26.4"
demucs = { git = "git@github.com:adefossez/demucs.git" }
torchaudio = "^2.1.0"
google-cloud-storage = "^2.14.0"
transformers = "^4.46.3"
google-cloud-aiplatform = "^1.41.0"
google-cloud = "^0.34.0"
google-cloud-container = "^2.40.0"
click = "^8.1.0"
soundfile = "^0.12.1"
lightning = "^2.2.0.post0"
datasets = "^2.17.1"
gcsfs = "^2023.10.0"
google-cloud-pubsub = "^2.19.6"
pytorch-metric-learning = "^2.4.1"
tensorboard = "^2.16.2"
levenshtein = "^0.25.1"
torch = "^2.1.0"
torchvision = "0.16.2"
evaluate = "^0.4.3"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
ipython = "^8.21.0"
python-dotenv = "^1.0.1"
ruff = "^0.2.1"
google-cloud-artifact-registry = "^1.11.5"
seaborn = "^0.13.2"
jiwer = "^3.0.4"
nltk = "^3.9.1"
sacrebleu = "^2.4.3"
rouge-score = "^0.1.2"
bert-score = "^0.3.13"
pytest = "^8.3.3"
google-cloud-firestore = "^2.19.0"
plotly = "^5.24.1"
openai = "^1.55.0"

[tool.poetry.group.zenml.dependencies]
kfp = "^2.9.0"
google-cloud-secret-manager = "^2.20.2"
google-cloud-container = "^2.51.0"
google-cloud-storage = "^2.18.2"
google-cloud-aiplatform = "^1.66.0"
google-cloud-build = "^3.24.2"
pygithub = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120


[tool.ruff]
line-length = 120
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # PyFlakes
    "B",      # flake8-bugbear
    "RUF100", # "noqa" rule ignore validation
    "I",
]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".ipynb",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.ruff.pydocstyle]
convention = "google"
